# Optimize semaphore (no locking, no allocations)