### Description

This PR optimizes the semaphore. The semaphore is on the hot path since every message needs to acquire it, an optimization will directly affect the performance of the pipeline. The changes include:
- The semaphore uses a `sync.Pool` to create and reuse channels, this reduces allocations to 0 in most cases.
- `Acquire` now returns a `Lock` that needs to be passed to `Release` instead of `Ticket`. This makes it impossible to release an unacquired ticket and makes the interface simpler since we don't need to return an error anymore.
- If the semaphore is used in an unexpected way (e.g. tickets/locks are acquired/released more than once) it will produce a panic.
- `Enqueue` no longer locks a mutex which improves performance, leaving it up to the caller to provide concurrency safety.

Benchmarks before change:
```
goos: darwin
goarch: amd64
pkg: github.com/conduitio/conduit/pkg/foundation/semaphore
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkNewSem-16             	1000000000	         0.2380 ns/op	       0 B/op	       0 allocs/op
BenchmarkEnqueueOneByOne/ticket-count-1-16         	12713606	        87.25 ns/op	      96 B/op	       1 allocs/op
BenchmarkEnqueueOneByOne/ticket-count-2-16         	 7086849	       169.1 ns/op	     192 B/op	       2 allocs/op
BenchmarkEnqueueOneByOne/ticket-count-8-16         	 1748378	       669.3 ns/op	     768 B/op	       8 allocs/op
BenchmarkEnqueueOneByOne/ticket-count-64-16        	  198808	      5517 ns/op	    6144 B/op	      64 allocs/op
BenchmarkEnqueueOneByOne/ticket-count-128-16       	  103171	     10977 ns/op	   12288 B/op	     128 allocs/op
BenchmarkEnqueueOneByOne/ticket-count-1024-16      	   13952	     86856 ns/op	   98304 B/op	    1024 allocs/op
BenchmarkEnqueueOneByOne/ticket-count-131072-16    	      98	  11110538 ns/op	12582924 B/op	  131072 allocs/op
BenchmarkEnqueueAll/ticket-count-1-16              	12632812	        88.25 ns/op	      96 B/op	       1 allocs/op
BenchmarkEnqueueAll/ticket-count-2-16              	 6957091	       176.4 ns/op	     192 B/op	       2 allocs/op
BenchmarkEnqueueAll/ticket-count-8-16              	 1687408	       702.0 ns/op	     768 B/op	       8 allocs/op
BenchmarkEnqueueAll/ticket-count-64-16             	  196902	      5641 ns/op	    6144 B/op	      64 allocs/op
BenchmarkEnqueueAll/ticket-count-128-16            	  101689	     11182 ns/op	   12288 B/op	     128 allocs/op
BenchmarkEnqueueAll/ticket-count-1024-16           	   12993	     88232 ns/op	   98304 B/op	    1024 allocs/op
BenchmarkEnqueueAll/ticket-count-131072-16         	      94	  12667224 ns/op	12582914 B/op	  131072 allocs/op
PASS
ok  	github.com/conduitio/conduit/pkg/foundation/semaphore	24.375s
```

Benchmarks after change:
```
goos: darwin
goarch: amd64
pkg: github.com/conduitio/conduit/pkg/foundation/semaphore
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkNewSem-16             	1000000000	         0.2368 ns/op	       0 B/op	       0 allocs/op
BenchmarkEnqueueOneByOne/ticket-count-1-16         	20424921	        55.76 ns/op	       0 B/op	       0 allocs/op
BenchmarkEnqueueOneByOne/ticket-count-2-16         	10444015	       109.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkEnqueueOneByOne/ticket-count-8-16         	 2701381	       446.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkEnqueueOneByOne/ticket-count-64-16        	  294818	      3543 ns/op	       0 B/op	       0 allocs/op
BenchmarkEnqueueOneByOne/ticket-count-128-16       	  158469	      7270 ns/op	       0 B/op	       0 allocs/op
BenchmarkEnqueueOneByOne/ticket-count-1024-16      	   20608	     57205 ns/op	       0 B/op	       0 allocs/op
BenchmarkEnqueueOneByOne/ticket-count-131072-16    	     164	   7143262 ns/op	      14 B/op	       0 allocs/op
BenchmarkEnqueueAll/ticket-count-1-16              	18595551	        57.00 ns/op	       0 B/op	       0 allocs/op
BenchmarkEnqueueAll/ticket-count-2-16              	 8788778	       131.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkEnqueueAll/ticket-count-8-16              	 2095458	       581.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkEnqueueAll/ticket-count-64-16             	  222130	      4741 ns/op	       0 B/op	       0 allocs/op
BenchmarkEnqueueAll/ticket-count-128-16            	  120446	      9638 ns/op	       0 B/op	       0 allocs/op
BenchmarkEnqueueAll/ticket-count-1024-16           	   15572	     76082 ns/op	       9 B/op	       0 allocs/op
BenchmarkEnqueueAll/ticket-count-131072-16         	     108	  10193201 ns/op	  174807 B/op	    1213 allocs/op
PASS
ok  	github.com/conduitio/conduit/pkg/foundation/semaphore	23.535s
```

Speed improvements by benchmark:
```
BenchmarkEnqueueOneByOne/ticket-count-1-16          36.09%
BenchmarkEnqueueOneByOne/ticket-count-2-16          35.19%
BenchmarkEnqueueOneByOne/ticket-count-8-16          33.27%
BenchmarkEnqueueOneByOne/ticket-count-64-16         35.78%
BenchmarkEnqueueOneByOne/ticket-count-128-16        33.77%
BenchmarkEnqueueOneByOne/ticket-count-1024-16       34.14%
BenchmarkEnqueueOneByOne/ticket-count-131072-16     35.71%
BenchmarkEnqueueAll/ticket-count-1-16               35.41%
BenchmarkEnqueueAll/ticket-count-2-16               25.28%
BenchmarkEnqueueAll/ticket-count-8-16               17.15%
BenchmarkEnqueueAll/ticket-count-64-16              15.95%
BenchmarkEnqueueAll/ticket-count-128-16             13.81%
BenchmarkEnqueueAll/ticket-count-1024-16            13.77%
BenchmarkEnqueueAll/ticket-count-131072-16          19.53%
```

### Quick checks:

- [X] I have followed the [Code Guidelines](https://github.com/ConduitIO/conduit/blob/main/docs/code_guidelines.md).
- [X] There is no other [pull request](https://github.com/ConduitIO/conduit/pulls) for the same update/change.
- [X] I have written unit tests.
- [X] I have made sure that the PR is of reasonable size and can be easily reviewed.